+++ History.md
@@ -1,3 +1,19 @@
+# 1.5.0 (2014-10-25)
+
+* capture all errors on remote socket and restart the tunnel
+
+# 1.4.0 (2014-08-31)
+
+* don't emit errors for ETIMEDOUT
+
+# 1.2.0 / 2014-04-28
+
+* return `client` from `localtunnel` API instantiation
+
+# 1.1.0 / 2014-02-24
+
+* add a host header transform to change the 'Host' header in requests
+
 # 1.0.0 / 2014-02-14
 
 * default to localltunnel.me for host